+++ xknx/cemi/cemi_handler.py
@@ -20,7 +20,7 @@ from xknx.exceptions import (
     DataSecureError,
     UnsupportedCEMIMessage,
 )
-from xknx.secure.data_secure import DataSecure
+from xknx.secure.data_secure import DataSecure, is_data_secure
 from xknx.secure.keyring import Keyring
 from xknx.telegram import IndividualAddress, Telegram, TelegramDirection, tpci
 from xknx.util import asyncio_timeout
@@ -119,8 +119,17 @@ class CEMIHandler:
             self.xknx.connection_manager.cemi_count_incoming_error += 1
             return
         logger.debug("Incoming CEMI: %s", cemi)
+        self.xknx.connection_manager.cemi_count_incoming += 1
 
-        if self.data_secure is not None:
+        if self.data_secure is None:
+            # if no keys are initialized, drop DataSecure frames
+            if is_data_secure(cemi.data):
+                data_secure_logger.debug(
+                    "Received DataSecure encrypted CEMI frame but no keys for DataSecure are initialized: %s",
+                    cemi,
+                )
+                return
+        else:
             try:
                 cemi.data = self.data_secure.received_cemi(cemi_data=cemi.data)
             except DataSecureError as err:
@@ -130,7 +139,7 @@ class CEMIHandler:
                     err,
                 )
                 return
-        self.xknx.connection_manager.cemi_count_incoming += 1
+
         telegram = cemi.data.telegram()
         telegram.direction = TelegramDirection.INCOMING
         self.telegram_received(telegram)